[MNG-6112] Central repository in the 4.0.0 super POM should declare update policy 'never'.
This reverts commit 13adc1bc2b6d9fc3a62ec2389996726c2edce4d8 *not changing
anything* just because someone did not react in time when this issue had been
lying around on a branch. If we would set pace for the industrie, everyone would
still be using 5.25" floppy disks and 8 bit hardware and we would tell them
that upgrading to be able to use new software would be too risky and they
better keep theire broken assembly routines until the machine goes up in
smoke.